LayoutTests/editing/selection/selection-toString-skipped-content-expected.txt

PASS getSelection().toString() is "hello"
PASS successfullyParsed is true

TEST COMPLETE
hello
